Renamed authsession to authSession
