+++ packages/identityprovider/src/validate-auth-session-id/scrim-user.ts
